+++ cmd/apply.go
@@ -72,6 +72,7 @@ var Apply = func() *cli.Command {
 					return fmt.Errorf("unable to create run")
 				}
 				fmt.Fprintf(os.Stdout, "created apply run %q\n", newRun.Uid)
+				fmt.Fprintln(os.Stdout, runs.GetBrowserUrl(cfg, workspace, *newRun))
 
 				if c.IsSet("wait") {
 					return runs.StreamLogs(ctx, cfg, workspace, newRun)